Seperated nft names and slugs into common javascript file
